% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-export.R
\name{connection}
\alias{connection}
\alias{connect}
\alias{disconnect}
\title{Database}
\usage{
connect(state = c("dev", "prod"))

disconnect(con)
}
\arguments{
\item{state}{Whether to connect to dev or prod.}

\item{con}{Connection, as returned by \code{connect}.}
}
\description{
Manage database connections for covid.
}
