% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawl.R
\name{crawl_covid}
\alias{crawl_covid}
\title{Crawl data}
\usage{
crawl_covid(
  puppet_ch = system.file("crawl/crawl_ch.js", package = "covid.crawler"),
  puppet_uk = system.file("crawl/crawl_uk.js", package = "covid.crawler"),
  date_time = Sys.time(),
  state = c("dev", "prod")
)
}
\arguments{
\item{puppet_ch, puppet_uk}{Paths to puppet js crawling files for UK and Switzerland.}

\item{date_time}{Date time of last update (dislpayed on dashboard).}

\item{state}{Where to push data to, dev of prod database.}
}
\description{
A wrapper functions for all \code{crawlers}.
}
